[XPU][P/D] Add XPU support to Nixl connector
