+++ vllm/distributed/kv_transfer/kv_connector/v1/nixl_connector.py
@@ -60,6 +60,7 @@ except ImportError:
 _NIXL_SUPPORTED_XPUS = {
     "cuda": ("cuda", ),
     "tpu": ("cpu", ),
+    "xpu": ("cpu", ),
 }
